+++ src/fortran/driver.f90
@@ -1,4 +1,4 @@
-PROGRAM lat_id_driver
+PROGRAM driver
   use find_kgrids, only: grid_selection, find_grids
   use kpointgeneration, only: generateIrredKpointList, mapKptsIntoFirstBZ
   use symmetry, only: bring_into_cell
@@ -8,7 +8,7 @@ PROGRAM lat_id_driver
   implicit none
 
   real(dp) :: lat_vecs(3,3), grid(3,3), offset(3), r_vecs(3,3), point(3), Rinv(3,3)
-  real(dp), allocatable :: grids(:,:,:)
+  real(dp), allocatable :: grids(:,:,:), offsets(:,:)
   real(dp) :: eps
   integer :: kpd, i
   real(dp), pointer :: IRKps(:,:)
@@ -28,16 +28,17 @@ PROGRAM lat_id_driver
 
   open(2,file="kpgen",status="old")
   read(2,*) kpd
-  read(2,*) offset
+  ! read(2,*) offset
   close(2)
 
   call matrix_inverse(transpose(lat_vecs),r_vecs)
   
-  call find_grids(lat_vecs,kpd,grids)
-  call grid_selection(grids, grid,r_vecs,offset,eps_=eps)
-
+  call find_grids(lat_vecs,kpd,grids,offsets,eps_=eps)
+  print *, "grids to search", size(grids,3)
+  call grid_selection(grids, grid, offset, r_vecs, offsets,eps_=eps)
+  
   call generateIrredKpointList(grid,r_vecs,offset,IRKps,weights,eps_=eps)
-  call mapKptsIntoFirstBZ(r_vecs,IRKps,eps_=eps)
+  ! call mapKptsIntoFirstBZ(r_vecs,IRKps,eps_=eps)
 
   open(4,file="KPOINTS")
   write(4,'(A29,I6.1,A14)') "Our new kpoint method. Found ", sum(weights), " total points."
@@ -48,7 +49,7 @@ PROGRAM lat_id_driver
      point = IRKps(i,:)
      call bring_into_cell(point,Rinv,r_vecs,eps)
      point = matmul(Rinv,IRKps(i,:))
-     write(4,'(F16.14,A1,F16.14,A1,F16.14,A1,I5.1)') point(1), " ",point(2), " ",point(3), " ", weights(i)
+     write(4,'(F16.14,A1,F16.14,A1,F16.14,A1,F5.1,A3,I5)') point(1), " ",point(2), " ",point(3), " ", real(weights(i)), " ! ", i
   end do
 
   
@@ -57,4 +58,4 @@ PROGRAM lat_id_driver
   !    print *, grid(:,i)
   ! end do
  
-end PROGRAM lat_id_driver
+end PROGRAM driver